ref(replays): hide selector widgets and disable tab if only mobile proj selected
